ev-range-Chatbot-project/requirements.txt
flask
flask-cors
openai
joblib
pandas
python-dotenv


